move tests for RocketCDN to Engine Directory/Namespace
